+++ contracts/interfaces/INFTLoanFacilitator.sol
@@ -97,7 +97,7 @@ interface INFTLoanFacilitator {
       * @dev only owner can call, value is scaled by SCALAR, 100% = SCALAR
       * @param feeRate the new origination fee rate
       */
-     event UpdateOriginationFeeRate(uint32 feeRate);
+     event UpdateOriginationFeeRate(uint256 feeRate);
 
      /**
       * @notice Emitted when requiredImprovementPercent is updated
@@ -116,13 +116,13 @@ interface INFTLoanFacilitator {
      * @notice The percent of the loan amount that the facilitator will take as a fee, scaled by SCALAR
      * @dev Starts set to 1%. Can only be set to 0 - 5%. 
      */
-    function originationFeeRate() external returns (uint32);
+    function originationFeeRate() external returns (uint256);
     
     /**
      * @notice The SCALAR for all percentages in the loan facilitator contract
      * @dev Any interest rate passed to a function should already been multiplied by SCALAR
      */
-    function SCALAR() external returns (uint40);
+    function SCALAR() external returns (uint256);
 
     /**
      * @notice The lend ticket contract associated with this loan faciliator
